change np.float64 to np.float32

+++ src/mlutils/sig.py
@@ -464,7 +464,7 @@ def generate_signatures(dbname, table,infile = None,outfile = None,models_folder
     
         with h5py.File(outfile, 'w') as hf:
             hf.create_dataset("keys", data = inchikeys[sort_idxs])
-            hf.create_dataset("V",  data = V[sort_idxs])
+            hf.create_dataset("V",  data = V[sort_idxs].astype(np.float32))
             hf.create_dataset("shape", data = V.shape)
     
         V = None
@@ -589,7 +589,7 @@ def generate_signatures(dbname, table,infile = None,outfile = None,models_folder
     
         with h5py.File(outfile, "w") as hf:
             hf.create_dataset("keys", data = inchikeys[sort_idxs])
-            hf.create_dataset("V", data = V[sort_idxs])
+            hf.create_dataset("V", data = V[sort_idxs].astype(np.float32))
             hf.create_dataset("shape", data = V[sort_idxs].shape)
         
         V = []
